Chapter 3 Exercise 8
Q: Write a program to test an integer value to determine if it is odd
or even. As always, make sure your output is clear and complete. In
other words, don't just put yes or no. Your output should stand alone
like 'The value 4 is an even number.' Hint: See the remainder (modulo)
operator in section 3.4.